fix: ping reply time

+++ mandatory/headers/icmp.h
@@ -6,7 +6,7 @@
 /*   By: root <root@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/19 19:01:28 by insub             #+#    #+#             */
-/*   Updated: 2026/01/26 03:00:14 by root             ###   ########.fr       */
+/*   Updated: 2026/01/26 03:29:30 by root             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 int		send_icmp_echo_request(int sockfd, const char *ip_addr);
 int		receive_icmp_echo_reply(int sockfd, char *buffer, int buf_size);
-void	process_icmp_reply(const char *buffer, int length, t_ping_stats *ping_stats, unsigned int ping_start_time_micro);
+void	process_icmp_reply(const char *buffer, int length, t_ping_stats *ping_stats, int64_t ping_start_time_micro);
 
 unsigned short	calculate_checksum(unsigned short *buf, int len);
